refactor: make content file to View file
